budget-killswitch/requirements.txt
google-cloud-run>=0.10.0
